// client/src/types/generated-message.ts
export interface GeneratedMessage {
  message: string;
  templateTitle: string;
  templateCategory: string;
  instruction: string;
  sourcePosts: string[];
  receiverName: string;
  receiverProfilePicture: string;
  receiverHeadline: string;
  senderName: string;
  senderProfilePicture: string;
  senderHeadline: string;
}